+++ hardpy/pytest_hardpy/utils/dialog_box.py
@@ -3,7 +3,7 @@
 
 from enum import Enum
 from dataclasses import dataclass
-from typing import Any
+from typing import Any, List, Optional, Union
 
 
 class DialogBoxWidgetType(Enum):
@@ -16,16 +16,26 @@ class DialogBoxWidgetType(Enum):
 
 
 @dataclass
+class RadiobuttonInfo:
+    fields: List[str]
+
+
+@dataclass
+class CheckboxInfo:
+    fields: List[str]
+
+
+@dataclass
 class DialogBoxWidget:
     """Dialog box widget.
 
     Args:
         type (DialogBoxWidgetType): widget type
-        info (dict | None): widget info
+        info (Union[RadiobuttonInfo, CheckboxInfo, None]): widget info
     """
 
     type: DialogBoxWidgetType
-    info: dict | None = None
+    info: Optional[Union[RadiobuttonInfo, CheckboxInfo]] = None
 
 
 @dataclass
@@ -58,6 +68,19 @@ def generate_dialog_box_dict(dialog_box_data: DialogBox) -> dict:
             "dialog_text": dialog_box_data.dialog_text,
             "widget": None,
         }
+    elif dialog_box_data.widget.info and dialog_box_data.widget.info.fields:
+        widget_info = dialog_box_data.widget.info
+        widget_data = {
+            "info": {
+                "fields": [str(field) for field in widget_info.fields],
+            },
+            "type": dialog_box_data.widget.type.value,
+        }
+        data_dict = {
+            "title_bar": dialog_box_data.title_bar,
+            "dialog_text": dialog_box_data.dialog_text,
+            "widget": widget_data,
+        }
     else:
         data_dict = {
             "title_bar": dialog_box_data.title_bar,
@@ -70,9 +93,7 @@ def generate_dialog_box_dict(dialog_box_data: DialogBox) -> dict:
     return data_dict
 
 
-def get_dialog_box_data(
-    input_data: str, widget: DialogBoxWidget | None
-) -> Any:
+def get_dialog_box_data(input_data: str, widget: DialogBoxWidget | None) -> Any:
     """Get the dialog box data in the correct format.
 
     Args:
@@ -81,6 +102,9 @@ def get_dialog_box_data(
 
     Returns:
         Any: Dialog box data in the correct format
+
+    Raises:
+        ValueError: If the widget.type is empty.
     """
     if widget is None:
         return None
@@ -88,13 +112,21 @@ def get_dialog_box_data(
     if widget.type is None:
         raise ValueError("Widget type is `None`, but widget data is not empty")
 
+    dbx_answer = None
+
     match widget.type:
         case DialogBoxWidgetType.NUMERIC_INPUT:
             try:
-                return float(input_data)
+                dbx_answer = float(input_data)
             except ValueError:
-                return None
+                dbx_answer = None
         case DialogBoxWidgetType.TEXT_INPUT:
-            return input_data
+            dbx_answer = input_data
+        case DialogBoxWidgetType.RADIOBUTTON:
+            dbx_answer = input_data
+        case DialogBoxWidgetType.CHECKBOX:
+            dbx_answer = input_data
         case _:
-            return None
+            pass  # noqa: WPS420
+
+    return dbx_answer